// src/app/modules/user/user.interface.ts
import { Model } from 'mongoose';
import { USER_ROLES } from '../../../enums/user';

export type IUser = {
  name: string;
  role: USER_ROLES;
  // contact: string;
  email: string;
  password: string;
  location: string;
  profile?: string;
  status: 'active' | 'delete';
  // verified: boolean;
  expertise?:string,
  authentication?: {
    isResetPassword: boolean;
    oneTimeCode: number;
    expireAt: Date;
  };
};

export type UserModal = {
  isExistUserById(id: string): any;
  isExistUserByEmail(email: string): any;
  isMatchPassword(password: string, hashPassword: string): boolean;
} & Model<IUser>;